Divided modal-logic.jl between types and utils folders
